refactor: entityPk column

See merge request s11-webmobile1-sub2/S11P12A302!7

+++ back/src/main/java/com/ssafy/algoFarm/solution/repository/AlgorithmSolutionRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.algoFarm.solution.repository;
 import com.ssafy.algoFarm.solution.entity.AlgorithmSolution;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlgorithmSolutionRepository extends JpaRepository<AlgorithmSolution, Integer> {
+public interface AlgorithmSolutionRepository extends JpaRepository<AlgorithmSolution, Long> {
 }
